refactor: Update AttributesComponent to export entityName, level, and entityType
This commit updates the AttributesComponent to export the entityName, level, and entityType properties. The entityName is set to "Villager" by default, the level is set to 1 by default, and the entityType is set to CharacterTypes.Type.NPC by default. This change allows for easier customization and configuration of character attributes.